[MIG] l10n_fr_department_delivery to v10
