+++ l10n_fr_department_delivery/README.rst
@@ -1,17 +1,33 @@
+.. image:: https://img.shields.io/badge/license-AGPL--3-blue.png
+   :target: https://www.gnu.org/licenses/agpl
+   :alt: License: AGPL-3
+
+=============================
 French Departments - Delivery
 =============================
 
-This module adds a field *French Departments* on delivery grids.
+This module adds a field *French Departments* on delivery carriers.
 
 Configuration
 =============
 
-To configure this module, go to *Warehouse > Configuration > Delivery > Delivery Pricelist* and, in the *Destination* tab, you will find the field *French Departments*.
+To configure this module, go to *Sales > Configuration > Delivery > Delivery Methods* and, in the *Destination* tab, you will find the field *French Departments*.
 
-Known issues
-============
+Usage
+=====
 
- * This module is not compatible with other modules that inherit the method *grid_get()* of the object *delivery.carrier*, which is the method that selects the delivery grid.
+.. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
+   :alt: Try me on Runbot
+   :target: https://runbot.odoo-community.org/runbot/121/10.0
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues
+<https://github.com/OCA/l10n-france/issues>`_. In case of trouble, please
+check there if your issue has already been reported. If you spotted it first,
+help us smashing it by providing a detailed and welcomed feedback.
+
 
 Credits
 =======
@@ -24,12 +40,14 @@ Contributors
 Maintainer
 ----------
 
-.. image:: http://odoo-community.org/logo.png
+.. image:: https://odoo-community.org/logo.png
    :alt: Odoo Community Association
-   :target: http://odoo-community.org
+   :target: https://odoo-community.org
 
 This module is maintained by the OCA.
 
-OCA, or the Odoo Community Association, is a nonprofit organization whose mission is to support the collaborative development of Odoo features and promote its widespread use.
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
 
-To contribute to this module, please visit http://odoo-community.org.
+To contribute to this module, please visit https://odoo-community.org.
