+++ l10n_fr_department_delivery/__manifest__.py
@@ -1,34 +1,17 @@
-# -*- encoding: utf-8 -*-
-##############################################################################
-#
-#    l10n FR Departments Delivery module for Odoo
+# -*- coding: utf-8 -*-
 #    Copyright (C) 2015 Akretion (http://www.akretion.com)
 #    @author Alexis de Lattre <alexis.delattre@akretion.com>
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 {
     'name': 'French Departments - Delivery',
     'summary': 'Use French Departments in delivery costs',
-    'version': '8.0.0.1.0',
+    'version': '10.0.0.1.0',
     'category': 'French Localization',
     'author': "Akretion,Odoo Community Association (OCA)",
     'website': 'http://www.akretion.com',
     'license': 'AGPL-3',
     'depends': ['l10n_fr_department', 'delivery'],
-    'data': ['view/delivery.xml'],
-    'installable': False,
+    'data': ['views/delivery.xml'],
+    'installable': True,
 }
