+++ l10n_fr_department_delivery/models/__init__.py
