--- l10n_fr_department_delivery/view/delivery.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-    Copyright (C) 2015 Akretion (http://www.akretion.com/)
-    @author: Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __openerp__.py
--->
-
-<openerp>
-<data>
-
-<record id="view_delivery_grid_form" model="ir.ui.view">
-    <field name="name">fr.department.delivery.grid.form</field>
-    <field name="model">delivery.grid</field>
-    <field name="inherit_id" ref="delivery.view_delivery_grid_form"/>
-    <field name="arch" type="xml">
-        <group string="States" position="after">
-            <group name="departments" string="French Departments">
-                <field name="department_ids" nolabel="1"/>
-            </group>
-        </group>
-    </field>
-</record>
-
-</data>
-</openerp>
